Move file to domain
