+++ src/domain/authorize.ts
@@ -1,7 +1,7 @@
 import * as E from 'fp-ts/lib/Either';
 import { pipe } from 'fp-ts/lib/function';
 import { ApiKey } from '../generated/definitions/ApiKey';
-import { Response } from '../domain/types';
+import { Response } from './types';
 
 export const authorizeApiKey = (apiKey: ApiKey): E.Either<Response<401>, ApiKey> =>
   pipe(
